Add a few tests for FirestoreDocument and fix odd behaviour of getCreatedTime
I wrote a few tests for FirestoreDocument, it's not complete but it's
what I had time for. In the course of testing I found that getCreatedTime
and getUpdatedTime was returning _now_ when no value was provided (such as
listing a non-existant document or creating a new document), so I changed
that as well.